any-py/backend/requirements.txt
# Flask and web framework dependencies
Flask==3.0.0
flask-cors==4.0.0

# MongoDB dependencies
pymongo==4.6.1

# Google Gemini AI
google-generativeai==0.8.0

# PDF generation
reportlab==4.0.9

# Security and utilities
Werkzeug==3.0.1
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Additional utilities
requests==2.31.0
typing-extensions==4.12.2